requirements.txt
appdirs==1.4.0
Django==1.10.5
django-ckeditor==5.2.1
django-parler==1.7
packaging==16.8
pyparsing==2.1.10
six==1.10.0
